Login Pages Security changes

+++ src/main/java/com/ouremr/product/repositories/PatientRegistrationRepository.java
@@ -4,5 +4,5 @@ import com.ouremr.product.tables.PatientRegistration;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
-public interface PatientRegistrationRepository extends JpaRepository<PatientRegistration, Integer>, JpaSpecificationExecutor<PatientRegistration> {
+public interface PatientRegistrationRepository extends JpaRepository<PatientRegistration, Long>, JpaSpecificationExecutor<PatientRegistration> {
 }
