+++ src/main/java/com/ouremr/product/repositories/SchedulerAppointmentRepository.java
@@ -4,5 +4,5 @@ import com.ouremr.product.tables.SchedulerAppointment;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
-public interface SchedulerAppointmentRepository extends JpaRepository<SchedulerAppointment, Integer>, JpaSpecificationExecutor<SchedulerAppointment> {
+public interface SchedulerAppointmentRepository extends JpaRepository<SchedulerAppointment, Long>, JpaSpecificationExecutor<SchedulerAppointment> {
 }
