+++ src/main/java/com/ouremr/product/repositories/SchedulerAppointmentStatusRepository.java
@@ -4,5 +4,5 @@ import com.ouremr.product.tables.SchedulerAppointmentStatus;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
-public interface SchedulerAppointmentStatusRepository extends JpaRepository<SchedulerAppointmentStatus, Integer>, JpaSpecificationExecutor<SchedulerAppointmentStatus> {
+public interface SchedulerAppointmentStatusRepository extends JpaRepository<SchedulerAppointmentStatus, Long>, JpaSpecificationExecutor<SchedulerAppointmentStatus> {
 }
