+++ src/main/java/com/ouremr/product/repositories/SchedulerAppointmentTypesRepository.java
@@ -4,5 +4,5 @@ import com.ouremr.product.tables.SchedulerAppointmentType;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
-public interface SchedulerAppointmentTypesRepository extends JpaRepository<SchedulerAppointmentType, Integer>, JpaSpecificationExecutor<SchedulerAppointmentType> {
+public interface SchedulerAppointmentTypesRepository extends JpaRepository<SchedulerAppointmentType, Long>, JpaSpecificationExecutor<SchedulerAppointmentType> {
 }
